requirements.txt
fastmcp>=2.3.0
mcp>=1.20.0
httpx>=0.28.0
python-dotenv>=1.0.0
uvicorn>=0.38.0
PyJWT>=2.10.0
Authlib>=1.6.0
starlette>=0.50.0
